bulk delete

+++ app/Livewire/Employee.php
@@ -17,6 +17,7 @@ class Employee extends Component
     public $updateData = false;
     public $employee_id;
     public $katakunci;
+    public $employee_selected_id = [];
 
     public function store()
     {
@@ -77,12 +78,23 @@ class Employee extends Component
         $this->alamat = '';
         $this->updateData = false;
         $this->employee_id = '';
+        $this->employee_selected_id = [];
     }
 
     public function delete()
     {
-        $id = $this->employee_id;
-        ModelsEmployee::find($id)->delete();
+
+        if ($this->employee_id != '') {
+            $id = $this->employee_id;
+            ModelsEmployee::find($id)->delete();
+        }
+        if (count($this->employee_selected_id)) {
+            for($x=0; $x < count($this->employee_selected_id) ; $x++){
+                ModelsEmployee::find($this->employee_selected_id[$x])->delete();
+            }
+        }
+
+        
         session()->flash('message','data berhasil didelete');
 
         $this->clear();
@@ -90,7 +102,9 @@ class Employee extends Component
 
     public function delete_confirmation($id)
     {
-        $this->employee_id = $id;
+        if ($id != '') {
+            $this->employee_id = $id;
+        }
     }
 
     public function render()
